refactor: рефактор страницы Books по замечаниям
